Fix swapped names
